requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==23.0.0
sentence-transformers==3.1.1
faiss-cpu==1.8.0
requests==2.32.3
pymongo==4.8.0
python-dotenv==1.0.1
langchain==0.3.0
pydantic==2.9.2
redis==5.0.1
fastapi-limiter==0.1.6